py-services/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
paddlepaddle>=2.5.0
paddleocr>=2.6.0
numpy>=1.20.0
pillow>=9.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0